Update Serializer object to be an interface (#219)
* Adding non serializable metadata handling

* removed an unused import

* update feature

* update changelog

* Add serializeObject public method

* Fixed linting

* fixed tests

* Implemented Serializer interface

* update scenarios

* update changelog

* Changes based on feedback

* update tests

* updates based on reviews

* Update bugsnag/src/main/java/com/bugsnag/Configuration.java


* update config

---------
