+++ bugsnag/src/main/java/com/bugsnag/serialization/Serializer.java
@@ -1,39 +1,17 @@
 package com.bugsnag.serialization;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import java.io.IOException;
 import java.io.OutputStream;
 
-public class Serializer {
-    private ObjectMapper mapper = new ObjectMapper();
-
+/**
+ * The Serializer is called to generate the JSON for an object to be added as metadata to a Bugsnag event.
+ */
+public interface Serializer {
     /**
-     * Constructor.
-     */
-    // Use deprecated method to ensure we don't break with older versions of jackson
-    @SuppressWarnings("deprecation")
-    public Serializer() {
-        mapper
-            .setSerializationInclusion(JsonInclude.Include.NON_EMPTY)
-            .setVisibilityChecker(
-                mapper.getVisibilityChecker().with(JsonAutoDetect.Visibility.NONE));
-    }
-
-    /**
-     * Write the object to the stream.
+     * Write the specified object to the provided stream to be used on metadata for a Bugsnag event.
      *
      * @param stream the stream to write the object to.
      * @param object the object to write to the stream.
      * @throws SerializationException the object could not be serialized.
      */
-    public void writeToStream(OutputStream stream, Object object) throws SerializationException {
-        try {
-            mapper.writeValue(stream, object);
-        } catch (IOException ex) {
-            throw new SerializationException("Exception during serialization", ex);
-        }
-    }
+    void writeToStream(OutputStream stream, Object object) throws SerializationException;
 }
